sr: separated oversized alloc ducktape tests as scale tests.
Moved the oversized alloc ducktape tests to be separate scale tests
and annotated them with @skip_debug_mode (since oversized alloc
warnings only occur in release). Added types for CI type checking.

(cherry picked from commit 2d1a114c109d3734555f68ea26d92acef1d4a62e)
